Convert project files from MSVC 2003 to MSVC 2005.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/4988


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@20776 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/samples/ufortune/ufortune.c
@@ -1,6 +1,6 @@
 /**************************************************************************
 *
-*   Copyright (C) 2001-2005, International Business Machines
+*   Copyright (C) 2001-2006, International Business Machines
 *   Corporation and others.  All Rights Reserved.
 *
 ***************************************************************************
@@ -187,7 +187,7 @@ int main(int argc, char **argv)
         exit(-1);
     }
 
-    i = time(NULL) % numFortunes;    /*  Use time to pick a somewhat-random fortune.  */
+    i = (int)time(NULL) % numFortunes;    /*  Use time to pick a somewhat-random fortune.  */
     resString = ures_getStringByIndex(fortunes_r, i, &len, &err);
     if (U_FAILURE(err)) {
         fprintf(stderr, "%s: ures_getStringByIndex(%d) failed, %s\n", programName, i, u_errorName(err));
